# openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare System API (RBAC + JWT)
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
  /auth/login:
    post:
      summary: Authenticate and issue JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required: [username, password]
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  expires_in: { type: integer }
        '401': { description: Invalid credentials }
  /auth/escalate:
    post:
      summary: Emergency access escalation for Doctors
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
              required: [reason]
      responses:
        '200': { description: Elevated token issued }
        '403': { description: Only Doctors can escalate }
  /records/{patient_id}:
    get:
      summary: View medical records
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: patient_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Records
          content:
            application/json:
              schema:
                type: object
                properties:
                  patient_id: { type: string }
                  records: { type: array, items: { type: object } }
        '403': { description: Patients can only view own records }
  /records/{patient_id}/status:
    patch:
      summary: Update patient status
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: patient_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
              required: [status]
      responses:
        '204': { description: No Content }
  /records/{patient_id}/prescriptions:
    post:
      summary: Prescribe medication
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: patient_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                medication: { type: string }
                dosage: { type: string }
              required: [medication, dosage]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
  /users:
    post:
      summary: Create user (Admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                role:
                  type: string
                  enum: [Patient, Nurse, Doctor, Administrator]
                password: { type: string }
              required: [username, role]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  username: { type: string }
                  role: { type: string }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT